Extracting classes from dual table selector and modifying to use actions. It currently is broken
