+++ appPtcgRandomizer/src/main/java/redactedrice/ptcgr/randomizer/actions/LambdaAction.java
@@ -18,9 +18,19 @@ public class LambdaAction  extends Action {
 		super (category, name, description);
 		this.perform = perform;
 	}
+	
+	protected LambdaAction(LambdaAction toCopy) {
+		super(toCopy);
+		this.perform = toCopy.perform;
+	}
 
 	@Override
-	public void Perform(Rom rom) {
+	public Action copy() {
+		return new LambdaAction(this);
+	}
+
+	@Override
+	public void perform(Rom rom) {
 		perform.perform(rom);
 	}
 }
